refactor: Move fallback extensions to supported-extensions.ts
- Consolidate fallback extensions functionality into supported-extensions.ts
- Update imports in parser.ts and test files
- Remove separate fallback-extensions.ts file
- All functionality remains the same, just better organized
