+++ src/services/code-index/processors/__tests__/parser.vb.spec.ts
@@ -11,7 +11,7 @@ vi.mock("../../../../../packages/telemetry/src/TelemetryService", () => ({
 	},
 }))
 
-import { shouldUseFallbackChunking } from "../../shared/fallback-extensions"
+import { shouldUseFallbackChunking } from "../../shared/supported-extensions"
 
 describe("CodeParser - VB.NET and Fallback Extensions Support", () => {
 	let parser: CodeParser
